pyproject.toml
[project]
name = "deal-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "cloudscraper>=1.2.71",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
